+++ packages/stitches-builder/CHANGELOG.md
@@ -9,4 +9,4 @@
 ### Patch Changes
 
 - Updated dependencies [40fdf74]
-  - @mochi-css/stitches@0.1.0
+    - @mochi-css/stitches@0.1.0